Removed redundant helpers folder and moved clan_points_helper.py to helper instead
